% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdb_api_cast.R
\name{create_cast_url}
\alias{create_cast_url}
\title{Create the cast url}
\usage{
create_cast_url(api_key, movie_id)
}
\arguments{
\item{api_key}{Api key of tmdb}

\item{movie_id}{Movie id of interest}
}
\value{
url needed to pass to the api
}
\description{
Given the api key and movie id, create the url needed to pass to the api
}
