% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdb_api_movies.R
\name{create_movie_url}
\alias{create_movie_url}
\title{Create url for api request}
\usage{
create_movie_url(api_key, genre_ids, min_date, max_date, page_num)
}
\arguments{
\item{api_key}{Api key requested from the website}

\item{min_date}{Min release date in the format 'yyyy-mm-dd'}

\item{max_date}{Max release date in the format 'yyyy-mm-dd'}

\item{genres}{genres of interest can include: 'Action', 'Adventure', 'Animation',
'Comedy', 'Crime','Documentary', 'Drama', 'Family', 'Fantasy', 'History',
'Horror', 'Music', 'Mystery', 'Romance', 'Science Fiction', 'TV Movie', 'Thriller',
 'War', 'Western'}
}
\value{
url needed for api request as a single string
}
\description{
Form the url needed for the api request
}
