% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network_data.R
\name{create_network_df}
\alias{create_network_df}
\title{Pull information from TMDB API}
\usage{
create_network_df(api_key, genres, min_date, max_date,
  movie_request_lim = 100)
}
\arguments{
\item{api_key}{Api key requested from the website}

\item{genres}{genres of interest can include: 'Action', 'Adventure', 'Animation',
'Comedy', 'Crime','Documentary', 'Drama', 'Family', 'Fantasy', 'History',
'Horror', 'Music', 'Mystery', 'Romance', 'Science Fiction', 'TV Movie', 'Thriller',
 'War', 'Western'}

\item{min_date}{Min release date in the format 'yyyy-mm-dd'}

\item{max_date}{Max release date in the format 'yyyy-mm-dd'}

\item{movie_request_lim}{Limit on the number of movies to pull}
}
\value{
A data frame containing movie titles and associated cast & crew
}
\description{
Combine Api pull functions 'pull_movie_titles', and 'pull_cast' to create
a dataframe that a network can be built from
}
