% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_network_data.R
\name{create_node_labels}
\alias{create_node_labels}
\title{Create labels for each node (each node is a person)}
\usage{
create_node_labels(links_df, movie_df, id_df)
}
\arguments{
\item{links_df}{Pairwise name combinations with id's dataframe.}

\item{movie_df}{a dataframe of movie data. the output from the function create_network_df}
}
\value{
A dataframe containing information for node labels, including which group they belong to (number of movies)
}
\description{
Taking information for each person, create a node label to be plotted along with the network
}
