% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_network_data.R
\name{find_most_popular}
\alias{find_most_popular}
\title{Return the k most popular people}
\usage{
find_most_popular(pairwise_df, id_df, between_df, network_df, k)
}
\arguments{
\item{pairwise_df}{Pairwise name combinations with id's dataframe. This parameter is the output
from add_ids_to_names function}

\item{id_df}{The id dataframe containing a column of names, and a column of id's}

\item{between_df}{a dataframe containing a column of names, and a column of betweenness scores}

\item{network_df}{a dataframe of movie data. the output from the function create_network_df}
}
\value{
A dataframe containing information for node labels, including which group they belong to (number of movies)
}
\description{
Return the k most popular people based on connections and betweenness
}
