% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdb_api_movies.R
\name{load_genres}
\alias{load_genres}
\title{Create movie genres data frame}
\usage{
load_genres()
}
\value{
data frame of genres and respective genre id
}
\description{
Create dataframe of the genres available on tmdb
}
