% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdb_api_movies.R
\name{match_genre_id}
\alias{match_genre_id}
\title{Match genre name to id needed for tmdb api}
\usage{
match_genre_id(genres)
}
\arguments{
\item{genres}{genres of interest can include: 'Action', 'Adventure', 'Animation',
'Comedy', 'Crime','Documentary', 'Drama', 'Family', 'Fantasy', 'History',
'Horror', 'Music', 'Mystery', 'Romance', 'Science Fiction', 'TV Movie', 'Thriller',
 'War', 'Western'}
}
\value{
A string containing a comma separated string of genre ids; passable to the api
}
\description{
Match the genre string to the genre id needed for the api request
}
