% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_network_data.R
\name{name_combinations}
\alias{name_combinations}
\title{Name Combinations}
\usage{
name_combinations(name_vect)
}
\arguments{
\item{name_vect}{A vector of cast or crew names}
}
\value{
a list of cast/crew  name combinations
}
\description{
Takes a vector of names and returns each 2-way connection as a dataframe
}
