% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdb_api_cast.R
\name{pull_cast}
\alias{pull_cast}
\title{Pull information on the cast of a movie}
\usage{
pull_cast(api_key, movie_id)
}
\arguments{
\item{api_key}{Api key of tmdb}

\item{movie_id}{Movie id of interest}
}
\value{
Dataframe of the cast for the movie of interest
}
\description{
Use tmdb api to return the cast of a movie given the tmdb movie id
}
